QFXFB-888 = Make ReceiptPrinterMQTTInterface a non-Lambda component again.
Convert ReceiptPrinterMQTTInterface back from a Lambda component to a
recipe-based component. This means it will stay running instead of being
invoked when the device receives an MQTT message.

It turns out that Greengrass only supports triggering Lambda components
with messages when the topic name is static. We need the topic name to be
different for each device so they only receive the print jobs directed to
them.