QFXFB-888 = Fix hardcoded MQTT endpoint address.
The MQTT endpoint address used in ReceiptPrinterMQTTInterface and
ReceiptPrinterHTTPInterface was hardcoded to the one for my dev environment
and wouldn't have worked for anyone else.

Also, in deployment.yaml, override all of the components' config variables
and document how to set them for different environments. We'll likely need
to override most of the config variables for deployments to production.