QFXFB-888 = Remove the version number from file/dir names.
It's easier to manage versions by using branches and tags instead of having
the version numbers in the filenames and keeping a copy of every version.
The AWS docs tell you to use the project structure we were using before,
but the open-source AWS components don't actually use it themselves.

The Greengrass CLI and `aws greengrassv2` expect you to use it, so this
commit works around that in the deployment scripts.