QFXFB-888 = Upgrade health-reporting to aws-iot-device-sdk-js-v2.
I thought the components were overwriting each others' shadow data and that
we needed to use named shadows (instead of the "classic" shadow) to prevent
that. The v1 SDK doesn't support named shadows, so I upgraded to use v2,
but then realised that I was wrong about the components overwriting the
data (or at least wrong about the reason for it).

It's still useful to upgrade to v2, mainly because v1 is now unmaintained
except for security fixed, so I'm committing it anyway.